feat: add album table

+++ src/artist/artist.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Album } from 'src/album/entities/album.entity';
 import { DatabaseService } from 'src/database/database.service';
 import { Repository } from 'typeorm';
 import { CreateArtistDto } from './dto/create-artist.dto';
@@ -11,6 +12,8 @@ export class ArtistService {
   constructor(
     @InjectRepository(Artist)
     private artistsRepository: Repository<Artist>,
+    @InjectRepository(Album)
+    private albumsRepository: Repository<Album>,
   ) {}
 
   async create(createArtistDto: CreateArtistDto): Promise<Artist> {
@@ -42,9 +45,24 @@ export class ArtistService {
     return artist;
   }
 
-  async remove(id: Artist['id']) {
+  async remove(id: Artist['id']): Promise<void> {
+    const artist = await this.artistsRepository.findOne({
+      where: { id },
+      relations: ['albums'],
+    });
+    if (!artist) throw new NotFoundException('Artist not found');
+
+    console.log(`artist.service.ts - line: 52 ->> artist`, artist);
+
+    for (const album of artist.albums) {
+      album.artistId = null;
+      await this.albumsRepository.save(album);
+    }
+
     const resultDelete = await this.artistsRepository.delete(id);
     if (resultDelete.affected === 0)
       throw new NotFoundException('Artist not found');
+
+    // await this.albumsRepository.update({ artistId: id }, { artistId: null });
   }
 }
